Add soleta-dev-app recipe and image configuration
In this patch also adds a new DISTRO_FEATURE called "soletta-dev-app".
This features install and configure systemd to autostart soletta-dev-app
on boot. This feature is disabled by default and to be enabled, it is needed
to add DISTRO_FEATURE_append = " soletta-dev-app" in yocto local.conf.

Now it is necessary to build using meta-oe since this layers contains nodejs
and git recipes.

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>