soletta-dev-app: use source-only repository for baking
To make soletta-dev-app installed by this recipe functional on
other architectures than x86_64, we need to build node modules
from scratch.

To do this, move to use the original source-only repository
that comes with nodejs packaging information and let bitbake
(using npm install) to fetch and build the necessary nodejs
module.

do_compile() is the courtesy of Sudarsana Nagineni.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>
Signed-off-by: Sudarsana Nagineni <sudarsana.nagineni@intel.com>