Revert "libpcre: properly enable utf8"
oe-core master (commit 839eebceecf33d106592bab154481486533ece75)
enabled unicode-properties by default in libpcre.

Because of that change and the fact that meta-soletta's bbappend
for libpcre has --enable-utf8 we start to get a configure error
(both --enable-utf and --enable-utf8 are enabled).

Instead of updating the bbappend to fix the error, we can just
remove it since it's no longer needed.

This reverts commit 60d7a8cee19723f986b57d9fb0f2d315b7d00809.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>