soletta-dev-app: Use git submodule fetcher
Use git submodule fetcher instead of calling git submodule manually.
This fixes a phase error (downloading the sources should happen in fetch
task instead of configure task) and helps with proxies.

Also, the fetcher does not need to check the remote if the
tag has changed and whether the local cache is good.

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>