soletta: allow custom custom path for libdir
Soletta's build system does not recognize the default flag for ${libdir} from
yocto as an environment variable to be used in the build, so it needs to be
exported manually.

This patch also removes hardcoded path that uses /usr/lib and replace it as
${libdir}.

This patch allows Soletta to be built in custom libdir paths
such as /usr/lib64 and any other valid path.

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>