soletta: remove INHIBIT_PACKAGE_STRIP and INHIBIT_DEFAULT_DEPS
The INHIBIT_PACKAGE_STRIP causes the build to not strip Soletta's binaries and
this is not the desired behavior.

The INHIBIT_DEFAULT_DEPS prevents the default dependencies, namely the C
compiler and standard C library (libc), from being added to DEPENDS and also
this is not the desired behavior for the recipe.

Signed-off-by: Bruno Bottazzini <bruno.bottazzini@intel.com>