soletta_git.bb: packaging compatible with OE-core master
Without this change, compilation with OE-core master fails with:

ERROR: soletta-1_beta19+gitAUTOINC+6a7bc32376-r0 do_package: QA Issue: soletta: Files/directories were installed but not shipped in any package:
   /usr/src
   /usr/src/debug
   /usr/src/debug/soletta
   /usr/src/debug/soletta/1_beta19+gitAUTOINC+6a7bc32376-r0
   ...

That's because of a packaging change in OE-core 0df6dabdf0a61a.

As discussed in
https://github.com/solettaproject/meta-soletta/issues/88,
INHIBIT_PACKAGE_STRIP and INHIBIT_DEFAULT_DEPS, are not desired
anymore, but removing them without other changes still leads to
packaging problems.

Part of those problems seem to be that default packaging rules are
overwritten instead of extending them. So this change here takes the
approach of adding to the defaults.

However, further work may be needed: something causes various packages
to depend on the -dev package, which triggers the dev-deps QA check
which (at least in Ostro OS) is a fatal error. For now this check is
suppressed.

Fixes: meta-soletta/#88

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>